+++ functions/src/index.ts
@@ -75,9 +75,11 @@ import {
   swarmspacePluginStatus,
   swarmspacePluginCatalog,
 } from "./functions/swarmspaceRouter";
+import { newsDataInvoke } from "./functions/newsDataInvoke";
 import { updateUserModelConfig } from "./functions/updateUserModelConfig";
 
 export { swarmspaceRouter };
 export { swarmspacePluginStatus };
 export { swarmspacePluginCatalog };
+export { newsDataInvoke };
 export { updateUserModelConfig };